rename mode.js - form.js
